apps/backend/pyproject.toml
[project]
name = "property-research-backend"
version = "0.1.0"
description = "Property research API backend with FastAPI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "pydantic[email]>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "psycopg2-binary>=2.9.9",           # PostgreSQL adapter
    "python-dotenv>=1.0.0",
    "python-json-logger>=4.0.0",
    "sqlmodel>=0.0.31",
    "pwdlib[argon2]>=0.3.0",
    "googlemaps>=4.10.0",
    "bcrypt==4.3.0",
    "pandas>=2.3.3",
    "geopandas>=1.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",          # For testing
    "black>=23.12.1",
    "ruff>=0.1.11",
    "mypy>=1.8.0",
    "ipython>=8.19.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.eggs
  | \.mypy_cache
  | \.pytest_cache
  | __pycache__
  | alembic/versions
)/
'''

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
exclude = ["alembic/versions/*", ".venv/*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "-v --cov=app --cov-report=term-missing"
